fix(pr): cancel a queued or auto-merge request instead of only reporting it
Returning a nonzero exit told the operator the merge had not landed, and left
the request standing. `gh pr merge` enqueues when the required checks pass and
enables auto-merge when they do not — and `--not-ready-reason` reaches that
second path directly — so GitHub could complete the merge minutes or hours
later without the ready-state read, the feedback ledger, the briefing, or the
typed confirmation. That is the merge nobody approved, arriving late.

The wrapper now cancels the pending request with `--disable-auto` when the
post-merge read reports anything other than MERGED, and says so. A failed
cancellation is reported loudly rather than swallowed, naming the consequence:
until it is cancelled by hand, the pull request can still merge later without
any of the gates. Both paths are tested.

That pushed the orchestrator past the 600-line cap — caught by the suite's own
assertion, which is what it is for — so the GitHub calls and the resolution
built on them move into `scripts/pr/merge-pr-github.mjs`. Four modules now:
ordering, pure decisions, local side effects, remote calls. All are named in
the routing arm and in the cap assertion, and sit at 453, 399, 193 and 191
lines.

Found by the closeout review.
